:memo: Change Moved the AppErrorSnackbarEffect class package to common.
